docker/docker-rpa/app/requirements.txt
blinker
certifi
charset-normalizer
click
Flask
grpcio
idna
itsdangerous
Jinja2
logic2-automation
MarkupSafe
numpy
opencv-python
pip
protobuf
pyserial
requests
urllib3
Werkzeug